+++ lib/Braintree/Dispute.php
@@ -138,7 +138,13 @@ class Dispute extends Base
      * Adds text evidence to a dispute, given a dispute ID and content
      *
      * @param string $id
-     * @param string $contentOrRequest
+     * @param string|mixed $contentOrRequest If a string, $contentOrRequest is the text-based content for the dispute evidence.
+     * Alternatively, the second argument can also be an array containing:
+     *  string $content The text-based content for the dispute evidence, and
+     *  string $category The category for this piece of evidence
+     *  Note: (optional) string $tag parameter is deprecated, use $category instead.
+     *
+     *  Example: https://developers.braintreepayments.com/reference/request/dispute/add-text-evidence/php#submitting-categorized-evidence
      */
     public static function addTextEvidence($id, $contentOrRequest)
     {