+++ tests/integration/PaymentMethodTest.php
@@ -473,7 +473,8 @@ class PaymentMethodTest extends Setup
 
         $this->assertFalse($result->success);
         $errors = $result->errors->forKey('paypalAccount')->errors;
-        $this->assertEquals(Braintree\Error\Codes::PAYPAL_ACCOUNT_CANNOT_VAULT_ONE_TIME_USE_PAYPAL_ACCOUNT, $errors[0]->code);
+        $this->assertContains(Braintree\Error\Codes::PAYPAL_ACCOUNT_CANNOT_VAULT_ONE_TIME_USE_PAYPAL_ACCOUNT,
+            array_map(function ($error) { return $error->code; }, $errors));
     }
 
     public function testCreate_handlesValidationErrorsForPayPalAccounts()
@@ -494,8 +495,10 @@ class PaymentMethodTest extends Setup
 
         $this->assertFalse($result->success);
         $errors = $result->errors->forKey('paypalAccount')->errors;
-        $this->assertEquals(Braintree\Error\Codes::PAYPAL_ACCOUNT_CANNOT_VAULT_ONE_TIME_USE_PAYPAL_ACCOUNT, $errors[0]->code);
-        $this->assertEquals(Braintree\Error\Codes::PAYPAL_ACCOUNT_CONSENT_CODE_OR_ACCESS_TOKEN_IS_REQUIRED, $errors[1]->code);
+        $this->assertContains(Braintree\Error\Codes::PAYPAL_ACCOUNT_CANNOT_VAULT_ONE_TIME_USE_PAYPAL_ACCOUNT,
+            array_map(function ($error) { return $error->code; }, $errors));
+        $this->assertContains(Braintree\Error\Codes::PAYPAL_ACCOUNT_CONSENT_CODE_OR_ACCESS_TOKEN_IS_REQUIRED,
+            array_map(function ($error) { return $error->code; }, $errors));
     }
 
     public function testCreate_allowsPassingDefaultOptionWithNonce()