+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## 5.4.0
+* Add `AcquirerReferenceNumber` to `Transaction`
+* Add `billingAgreementId` to `PayPalDetails`
+* Deprecate `recurring` in Transaction sale
+* Deprecate `tag` in Dispute add text evidence
+
 ## 5.3.1
-- Deprecate `masterpassCard` and `amexExpressCheckoutCard` payment methods
-- Deprecate `amexExpressCheckoutCardDetails`
+* Deprecate `masterpassCard` and `amexExpressCheckoutCard` payment methods
+* Deprecate `amexExpressCheckoutCardDetails`
 
 ## 5.3.0
 * Add `RISK_THRESHOLD` to `GatewayRejectionReason` constants